move gui logic for menu into new files gui.c/gui.h
