+++ src/menu_screen.c
@@ -1,7 +1,8 @@
 #include "screens.h"
 #include "common.h"
-#include "raylib.h"
-#include <stdlib.h> //malloc
+// #include "raylib.h"
+#include "gui.h"
+#include <stdlib.h> //malloc, NULL
 
 //TODO:: Move elsewhere, commons maybe
 #define SQUARE_PIXEL_WIDTH 16
@@ -24,13 +25,6 @@ typedef enum {
 } BackgroundSprite;
 
 typedef enum {
-    ALIGN_LEFT,
-    ALIGN_CENTER,
-    ALIGN_BELOW,
-    ALIGN_ABOVE
-} TextAlignment;
-
-typedef enum {
     MAP_SMALLSIZE,
     MAP_MEDIUMSIZE,
     MAP_LARGESIZE,
@@ -44,52 +38,6 @@ typedef enum {
     RES_2080x2080,
     RES_COUNT
 } ScreenRes;
-
-// typedef enum {
-//     MENU_DIRTBACKGROUND,
-//     MENU_WHITEBACKGROUND,
-//     MENU_BACKGROUNDCOUNT
-// } MapBackground;
-
-
-typedef struct UiElement{
-    bool draw_rect; //may be false for text only
-    bool use_texture;
-    Rectangle rect;
-    Color inner_color;
-    int border_thickness;
-    Color border_color;
-    Rectangle texture_rect;
-    Texture2D inner_texture;
-
-    bool draw_glow;
-    int glow_thickness;
-    Color glow_color;
-    
-    char* text;
-    Color text_color;
-    Font text_font;
-    float text_size;
-    float text_spacing;
-    TextAlignment text_align;
-} UiElement;
-
-typedef struct {
-    UiElement* boxes;
-    int count;
-    int selected; //none is -1
-    int hovering; //none is -1
-    Color hover_glow_color;
-    int hover_glow_thickness;
-    Color selected_glow_color;
-    int selected_glow_thickness;
-} UiBoxGroup;
-
-typedef struct {
-    int count;
-    int max;
-    UiElement* buffer;
-} ElementArena;
 
 typedef struct MenuGui{
     UiElement* mouse_overelem;
@@ -112,84 +60,11 @@ typedef struct MenuGui{
 } MenuGui;
 
 
-// typedef struct MenuScene{
-//     Camera2D camera;
-//     MenuGui menu_gui;
-//     Texture2D* textures;
-// } MenuScene;
-
-Rectangle GetSpriteRect(int sprite_index, int sprite_width, bool flip_x, bool flip_y);
-MenuGui*  setup_snakemenu(Texture2D* textures); //Tightly tied to game logic....
-void draw_menu(MenuGui* menu);
-void init_uielement(UiElement* element);
-void draw_uielement(UiElement* element); //Necessary to export?
-
-
 Rectangle GetSpriteRect(int sprite_index, int sprite_width, bool flip_x, bool flip_y) {
     Rectangle rect = {sprite_width*sprite_index, 0, sprite_width, sprite_width};
     rect.width = flip_x ? -1 * rect.width : rect.width;
     rect.height = flip_y ? -1 * rect.height : rect.height;
     return rect;
-}
-
-//set default values for UI element (zeroing where appropriate)
-//DOES NOT CONSTRUCT ELEMENT
-//Relies on GetFontDefault from raylib
-void init_uielement(UiElement* element) {
-    element->draw_rect = true;
-    element->use_texture = false;
-    element->rect = (Rectangle){0};
-    element->inner_color = WHITE;
-    element->border_thickness = 0;
-    element->border_color = BLACK;
-    element->texture_rect = (Rectangle){0};
-    element->inner_texture = (Texture2D){0};
-
-
-    element->draw_glow = false;
-    element->glow_thickness = 0;
-    element->glow_color = (Color){0};
-    
-    element->text = NULL;
-    element->text_color = BLACK;
-    element->text_font = GetFontDefault();
-    element->text_size = 12.0f;
-    element->text_spacing = 1.0f;
-    element->text_align = ALIGN_CENTER;
-}
-
-void elemarena_alloc(ElementArena* arena, int size) {
-    if (arena->buffer) {free(arena->buffer);}
-    arena->buffer = malloc(sizeof(UiElement) * size);
-    arena->count = 0;
-    arena->max = size;
-}
-
-void elemarena_dealloc(ElementArena* arena) {
-    if (arena->buffer) {
-        free(arena->buffer);
-    }
-    arena->count = 0;
-    arena->max = 0;
-}
-
-//Adds a number of elements to the GUI, returns pointer to first in the new range of elements.
-//If there's not enough room, arena buffer is reallocated.
-UiElement* elemarena_addelems(ElementArena* arena, int count) {
-    if (count + arena->count > arena->max) {
-        int newmax = (count + arena->count) * 1.5;
-        UiElement* newbuf = malloc(sizeof(UiElement) * newmax);
-        for (int i = 0; i < arena->count; i++) {
-            newbuf[i] = arena->buffer[i];
-        }
-        free(arena->buffer);
-        arena->buffer = newbuf;
-        arena->max = newmax;
-    }
-
-    UiElement* start = &arena->buffer[arena->count];
-    arena->count += count;
-    return start;
 }
 
 
@@ -199,6 +74,7 @@ void setup_menu(MenuGui* menu, GameState* state) {
     menu->mouse_overelem = NULL;
     menu->mouse_downelem = NULL;
     menu->clicked_elem = NULL;
+    menu->elem_arena.buffer = NULL;
 
     //Allocate elements
     elemarena_alloc(&menu->elem_arena, 15);
@@ -337,93 +213,7 @@ void setup_menu(MenuGui* menu, GameState* state) {
     menu->start_button->text_size = 30.0f;
 }
 
-//Draw UI element to current render target
-void draw_uielement(UiElement* element) {
-    if (element->draw_glow) {
-        DrawRectangleGradientV(element->rect.x,
-                    element->rect.y - element->glow_thickness,
-                    element->rect.width,
-                    element->glow_thickness,
-                    (Color){element->glow_color.r,element->glow_color.g,element->glow_color.b,0},
-                    element->glow_color);
-        DrawRectangleGradientV(element->rect.x,
-                    element->rect.y + element->rect.height,
-                    element->rect.width,
-                    element->glow_thickness,
-                    element->glow_color,
-                    (Color){element->glow_color.r,element->glow_color.g,element->glow_color.b,0});
-        DrawRectangleGradientH(element->rect.x - element->glow_thickness,
-                    element->rect.y,
-                    element->glow_thickness,
-                    element->rect.height,
-                    (Color){element->glow_color.r,element->glow_color.g,element->glow_color.b,0},
-                    element->glow_color);
-        DrawRectangleGradientH(element->rect.x + element->rect.width,
-                    element->rect.y,
-                    element->glow_thickness,
-                    element->rect.height,
-                    element->glow_color,
-                    (Color){element->glow_color.r,element->glow_color.g,element->glow_color.b,0});
-    }
 
-    
-    if (element->draw_rect) {
-        if (element->border_thickness) {
-            Rectangle border_rect = (Rectangle){
-                element->rect.x - element->border_thickness,
-                element->rect.y - element->border_thickness,
-                element->rect.width + element->border_thickness * 2,
-                element->rect.height + element->border_thickness * 2
-                };
-            DrawRectangleRec(border_rect, element->border_color);
-        }
-        if (element->use_texture) {
-            DrawTexturePro(element->inner_texture, element->texture_rect, element->rect, (Vector2){0,0}, 0.0f, WHITE);
-        } else {
-            DrawRectangleRec(element->rect, element->inner_color);
-        }
-
-    }
-    
-    if (element->text) {
-        Vector2 t_sz = MeasureTextEx(element->text_font, element->text, element->text_size, element->text_spacing);
-        Vector2 pos;
-        switch(element->text_align) {
-            case ALIGN_LEFT:
-                pos.x = element->rect.x;
-                pos.y = element->rect.y + element->rect.height/2 - t_sz.y / 2;
-                break;
-            case ALIGN_CENTER:
-                pos.x = element->rect.x + element->rect.width/2 - t_sz.x/2;
-                pos.y = element->rect.y + element->rect.height/2 - t_sz.y / 2;
-                break;
-            case ALIGN_ABOVE:
-                pos.x = element->rect.x + element->rect.width/2 - t_sz.x/2;
-                pos.y = element->rect.y - element->text_spacing - t_sz.y;
-                break;
-            case ALIGN_BELOW:
-                pos.x = element->rect.x + element->rect.width/2 - t_sz.x/2;
-                pos.y = element->rect.y + element->text_spacing + element->rect.height;
-                break;
-        }
-
-        DrawTextEx(element->text_font, 
-            element->text, 
-            pos, 
-            element->text_size, 
-            element->text_spacing, 
-            element->text_color);
-    }
-}
-
-
-void draw_uiboxgroup(UiBoxGroup* group) {
-    //handle glow effect based on selected
-    //handle glow effect based on hover
-    for (int i = 0; i < group->count; i++) {
-        draw_uielement(&group->boxes[i]);
-    }
-}
 
 
 void setup_menuscreen(GameState* state) {
@@ -438,12 +228,7 @@ void setup_menuscreen(GameState* state) {
     state->screen_memory = menu;
 }
 
-bool is_inelementbounds(UiElement* elem, Vector2 pos) {
-    return  elem->rect.x < pos.x &&
-            elem->rect.y < pos.y &&
-            elem->rect.x + elem->rect.width > pos.x &&
-            elem->rect.y + elem->rect.height > pos.y;
-}
+
 
 void update_menuscreen(GameState* state) {
     MenuGui* menu = state->screen_memory;
@@ -462,7 +247,7 @@ void update_menuscreen(GameState* state) {
             menu->clicked_elem = menu->mouse_overelem;
             menu->mouse_downelem = NULL;
         }
-        if (IsMouseButtonDown(MOUSE_BUTTON_LEFT) && menu->mouse_overelem) {
+        if (IsMouseButtonPressed(MOUSE_BUTTON_LEFT) && menu->mouse_overelem) {
             menu->mouse_downelem = menu->mouse_overelem;
         }
     }
